project
Hi,
I want to do a project in charge of transmitting information from sensors. The controller will receive the information, convert it with A / D, and transfer the information over the Internet to the computer in my home. Does anyone have any idea how I can start creating this project?
I know I need an A / D program and an Internet program, but how do I connect the programs?